digital_speedometer_widget: Use uint8_t for GPS speed as well

+++ src/user_interface/digital_speedometer_widget.cc
@@ -51,7 +51,9 @@ DigitalSpeedometerWidget::Update(ApplicationState& state)
     auto conf = ro.Get<AS::configuration>();
 
     lv_label_set_text(m_speed_digits_label, std::format("{}", ro.Get<AS::speed>()).c_str());
-    lv_label_set_text(m_gps_speed_label, std::format("{}", ro.Get<AS::position>()->speed).c_str());
+    lv_label_set_text(
+        m_gps_speed_label,
+        std::format("{}", static_cast<uint8_t>(ro.Get<AS::position>()->speed)).c_str());
 
     if (conf->speedometer_type == SpeedometerType::kDigital ||
         conf->speedometer_type == SpeedometerType::kBoth)
